LATIR-1: Refactor code
