+++ src/components/AppContainer/index.js
@@ -2,10 +2,14 @@ import React from 'react'
 import styled from 'styled-components'
 import Navbar from '../Navbar';
 
-const Container = styled.div`
+const Content = styled.main`
   width: 90%;
   height: 90%;
   margin: 0;
+  padding-top: 64px;
+`
+
+const Container = styled.div`
   font-family: QuinoaRound;
 `
 
@@ -13,7 +17,9 @@ const AppContainer = (props) => {
   return(
     <Container>
       <Navbar />
-      {props.children}
+      <Content>
+        {props.children}
+      </Content>
     </Container>
   )
 }
