+++ src/components/LogoIcon/index.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import Colors from '../../utils/colors';
+
+const LogoIcon = styled.div`
+    width: 40px;
+    height: 40px;
+    margin: 6px 12px;
+    background-color: ${Colors.Primary.hexString()};
+`
+
+export default LogoIcon;
