+++ src/components/NoSelect/index.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const NoSelect = styled.div`
+    user-select: none;
+`
+
+export default NoSelect
