+++ src/components/index.js
@@ -1,4 +1,5 @@
 // Here we should export all components for easy access
 export { default as AppContainer } from './AppContainer'; 
 export { default as HamburgerButton } from './HamburgerButton';
-export { default as Navbar } from './Navbar'; 
+export { default as Navbar } from './Navbar';
+export { default as NoSelect } from './NoSelect';
