+++ src/hooks/index.js
@@ -1,0 +1,8 @@
+export { 
+    default as useDevice ,
+    getDevice,
+    isLaptop,
+    isTablet,
+    isMobile
+} from './useDevice';
+export { default as useWindowSize } from './useDevice'
